requirements.txt
dj-database-url==0.4.1
Django==1.10

django-widget-tweaks==1.4.2

gcloud==0.17.0
googleapis-common-protos==1.5.3

Jinja2==2.10
jsonify==0.5

model-mommy==1.2.6
nodejs==0.1.1
oauth2client==3.0.0
oauthlib==2.1.0
olefile==0.44
optional-django==0.1.0
Pillow==4.3.0
protobuf==3.6.1
psycopg2==2.7.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycryptodome==3.4.3

social-auth-app-django==2.1.0
social-auth-core==1.7.0
pagseguro==0.3.4
django-paypal==0.4.1
South==1.0.2
requests==2.11.1
xmltodict==0.10.2
arrow==0.8.0
python-dateutil==2.7.3
google-api-python-client==1.4.1
django-jet==1.0.7
easy-thumbnails==2.3
uritemplate==3.0.0
django-widget-tweaks==1.4.3
django-watson==1.5.2
django-newsletter==0.7